src/data/usecases/schedule-appointment/remote-load-date-list.ts
import { LoadDateSchedule } from "src/domain/usecases";
import { NotFoundError, UnexpectedError } from "src/domain/errors";
import { HttpClient, HttpStatusCode } from "src/data/protocols/http";

export class RemoteLoadDateList implements LoadDateSchedule {
  constructor(
    private readonly url: string,
    private readonly httpClient: HttpClient<LoadDateSchedule.Model>
  ) {}

  async loadAll() {
    const httpResponse = await this.httpClient.request({
      url: this.url,
      method: "get",
    });

    switch (httpResponse.statusCode) {
      case HttpStatusCode.ok:
        if (!httpResponse.body) {
          throw new UnexpectedError();
        }

        return httpResponse.body;

      case HttpStatusCode.notFound:
        throw new NotFoundError();

      default:
        throw new UnexpectedError();
    }
  }
}
